pyproject.toml
[project]
name = "incident-tracker"
version = "0.1.0"
description = "sistema interno onde a equipe de operações e os desenvolvedores relatam quedas de servidores, bugs críticos em produção e gerenciam o status e a prioridade da resolução."
authors = [
    {name = "Vítor",email = "vitorhugo8899o@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "validate-docbr (>=1.11.1,<2.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.8,<0.16.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'